Add document ResourceMessage
