+++ editor/src/messages/portfolio/document/mod.rs
@@ -7,6 +7,7 @@ pub mod navigation;
 pub mod node_graph;
 pub mod overlays;
 pub mod properties_panel;
+pub mod resource;
 pub mod utility_types;
 
 #[doc(inline)]
