+++ editor/src/messages/portfolio/document/resource/resource_message.rs
@@ -1,0 +1,7 @@
+use crate::messages::prelude::*;
+
+#[impl_message(Message, DocumentMessage, Resource)]
+#[derive(PartialEq, Clone, Debug, serde::Serialize, serde::Deserialize)]
+pub enum ResourceMessage {
+	Noop,
+}
