+++ editor/src/messages/portfolio/document/utility_types/mod.rs
@@ -4,6 +4,5 @@ pub mod error;
 pub mod misc;
 pub mod network_interface;
 pub mod nodes;
-pub mod resources;
 pub mod transformation;
 pub mod wires;
